compressdev: modify size of max nb queue pairs
Number of queue pairs used everywhere in compressdev
is uint16_t, so maximum number of queue pairs should be
uin16_t too.

+++ lib/librte_compressdev/rte_compressdev.h
@@ -181,7 +181,7 @@ struct rte_compressdev_info {
 	uint64_t feature_flags;			/**< Feature flags */
 	const struct rte_compressdev_capabilities *capabilities;
 	/**< Array of devices supported capabilities */
-	unsigned int max_nb_queue_pairs;
+	uint16_t max_nb_queue_pairs;
 	/**< Maximum number of queues pairs supported by device.
 	 * (If 0, there is no limit in maximum number of queue pairs)
 	 */
